+++ eras/shelley/impl/CHANGELOG.md
@@ -24,6 +24,9 @@
 
 ### `testlib`
 
+* Add argument to `impSatisfyNativeScript` that accounts for already satisifed witnesses
+* Remove `impFreshIdxL`
+* Remove root coin argument from `initImpNES`, `initShelleyImpNES`
 * Adjust `sendCointTo` and `sendValueTo` to return `TxIn`
 * Added `ToExpr` instance for `ShelleyTxAuxDataRaw`
 * Add: